refactor: 136 - move slider into another component
