+++ src/components/catalog/layout.ts
@@ -3,7 +3,7 @@ import { createSearchWrapper } from './search';
 import { fetchProductCards } from '../../helpers/fetch-product-cards';
 import { CATALOG } from '../../pages/catalog/constants';
 import { createSortComponent } from './sort';
-import { createPriceFilter } from './filter-price';
+import { createPriceFilter } from './filter-price/filter-price';
 
 export async function catalogLayout(): Promise<HTMLDivElement> {
   const layout = createDiv({ classes: CATALOG.wrapper });
